requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.36
pymysql>=1.1.1
cryptography>=43.0.0
python-dotenv>=1.0.1
apscheduler>=3.10.4
httpx>=0.27.0
feedparser>=6.0.11
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
pydantic>=2.9.0
pydantic-settings>=2.5.0
google-auth>=2.35.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.150.0
requests>=2.31.0
